move alloc-size helpers to mem_cache.common; drop lazy imports
